Fix data error when Timelimit set to Partition_Limit
The Timelimit field can be set to the string Partition_Limit
in cases where a user cancels a job.  This is the value chosen
by slurm's job_info.c when time_limit has no value.

Update the slurmtime() function to detect this value and return
None if present.  This avoids a parse error for the Timelimit
column.

Fixes issue #3